fix set the profile for the tests to no_ssl explicitly
With SAML authentication incorporated many of the tests were failing with authentication issues.
Set the profile to 'no_ssl' in order for the requests from the integration tests to work.
The tests are setup with the user and the role information as part of the test where needed.  The tests will
not be able to perform a SAML authentication check.